update tests to write to StringIO
