+++ python_modules/dagster/dagster_tests/core_tests/log_tests/test_logging.conf
@@ -1,0 +1,42 @@
+[loggers]
+keys=root,loggerOne,loggerTwo
+
+[handlers]
+keys=consoleHandlerOne,consoleHandlerTwo
+
+[formatters]
+keys=simpleFormatterOne,simpleFormatterTwo
+
+[logger_root]
+level=DEBUG
+handlers=consoleHandlerOne
+
+[logger_loggerOne]
+level=DEBUG
+handlers=consoleHandlerOne
+qualname=userLogger
+propagate=0
+
+[logger_loggerTwo]
+level=DEBUG
+handlers=consoleHandlerTwo
+qualname=userLogger
+propagate=0
+
+[handler_consoleHandlerOne]
+class=StreamHandler
+level=DEBUG
+formatter=simpleFormatterOne
+args=(sys.stdout,)
+
+[handler_consoleHandlerTwo]
+class=StreamHandler
+level=DEBUG
+formatter=simpleFormatterTwo
+args=(sys.stdout,)
+
+[formatter_simpleFormatterOne]
+format=%(name)s :: %(message)s
+
+[formatter_simpleFormatterTwo]
+format=%(name)s :: %(message)s
