--- python_modules/libraries/dagster-postgres/logging.conf
@@ -1,27 +1,0 @@
-[loggers]
-keys=root,userLogger
-
-[handlers]
-keys=consoleHandler
-
-[formatters]
-keys=simpleFormatter
-
-[logger_root]
-level=DEBUG
-handlers=consoleHandler
-
-[logger_userLogger]
-level=DEBUG
-handlers=consoleHandler
-qualname=userLogger
-propagate=0
-
-[handler_consoleHandler]
-class=StreamHandler
-level=DEBUG
-formatter=simpleFormatter
-args=(sys.stdout,)
-
-[formatter_simpleFormatter]
-format=%(asctime)s :: %(name)s :: %(levelname)s :: %(message)s
